kubeflow_pipelines/pipeline/requirements_vertex_ai.txt
kfp==1.8.2
google-cloud-storage==1.42.2
google-cloud-secret-manager==2.7.1
